split utility function to different files according to relation and namespace
