+++ src/common/auto_fd.cpp
@@ -1,5 +1,5 @@
 #include "auto_fd.hpp"
-#include "utils.hpp"
+#include "os_utils.hpp"
 
 AutoFd::AutoFd(const int fd)
     : _fd(fd)
