+++ src/common/socket.cpp
@@ -3,10 +3,9 @@
 #include <poll.h>
 #include <unistd.h>
 
-#include "socket.hpp"
-#include "utils.hpp"
-#include "../exceptions.hpp"
 #include "server_socket.hpp"
+#include "socket.hpp"
+#include "os_utils.hpp"
 
 Socket::Socket(const int socket_fd)
     : _socket_fd(socket_fd)
