+++ src/common/string_utils.hpp
@@ -1,0 +1,14 @@
+#pragma once
+#include <string>
+#include <vector>
+
+namespace strings
+{
+    std::vector<std::string> split(const std::string &string, char token, size_t max);
+
+    std::string strip(const std::string &str, char token);
+
+    std::string to_string(char c);
+
+    bool starts_with(const std::string &str, const std::string &prefix);
+};
