+++ src/common/temp_folder.cpp
@@ -2,7 +2,7 @@
 using namespace std::filesystem;
 
 #include "temp_folder.hpp"
-#include "utils.hpp"
+#include "os_utils.hpp"
 
 os::TempFolder::TempFolder(const std::string &template_name)
     : _path(template_name)
