+++ src/rules_utils.hpp
@@ -1,0 +1,7 @@
+#include "proto/rule.pb.h"
+#include "block_device.hpp"
+
+namespace rules
+{
+    bool is_rule_matching(const RuleFilter &rule, const BlockDevice &device);
+};
